LPS-126713 Moves to journal.article package
